refactor(dnd): move resolveDndCapabilities out of the public barrel
`resolveDndCapabilities` and `DndCapabilities` are implementation
details — they map the user-facing `dndStrategy` option onto per-backend
flags the construction sites consume. Consumers should only ever see
`dndStrategy`. Both were inadvertently public via `export * from
'./dockview/options'`.

Moved to a new internal `src/dockview/dndCapabilities.ts` not in the
package barrel. `DockviewDndStrategy` stays in `options.ts` (and remains
exported) since it names the `dndStrategy` option's value type.

937/937 pass; full monorepo build green.
